Switched to a more consistent code style
My atomic `RUN` command was not bad, but there already was a command to create the home directory earlier in the file, and there also was a section regarding the `sshd` config, which used `sed` instead of `printf`.
With a comment explaining my process, I do hope this is clear and more consistent with the style of the Dockerfile.
Thanks to dduportal for his help and patience.